a couple of edge cases fixed and READ.me added

+++ src/Order.h
@@ -12,8 +12,8 @@ public:
     Side Side;
     double Price;
     int Quantity;
-    int OriginalQuantity;
-    long long Sequence;
+    int OriginalQuantity; // so that even if quanity is updated the original is kept
+    long Sequence; // to see which order arrived first. (rather than including this in the data structure level)
 
     Order()
         : Side(Side::Buy), Price(0.0), Quantity(0), OriginalQuantity(0), Sequence(0) {}
